refactor(plot_domain): rename to realisations.py and extract more utility functions
